requirements.txt
duckdb==1.3.1
jupyterlab==4.4.4
jupyterlab_server==2.27.3
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.1
scikit-learn==1.7.0
manimgl==1.7.2
pytesseract==0.3.13
pillow==11.3.0
ollama==0.5.1
python-dotenv==1.1.1